+++ src/controllers/index.js
@@ -1,6 +1,6 @@
-import userController from './userController.js';
-import productController from './productController.js';
-import orderController from './orderController.js';
-import authController from './authController.js';
+import userController from './user-controller.js';
+import productController from './product-controller.js';
+import orderController from './order-controller.js';
+import authController from './auth-controller.js';
 
 export { userController, productController, orderController, authController };
