+++ src/db/models/index.js
@@ -1,5 +1,5 @@
-import userModel from './userModel.js';
-import productModel from './productModel.js';
-import orderModel from './orderModel.js';
+import userModel from './user-model.js';
+import productModel from './product-model.js';
+import orderModel from './order-model.js';
 
 export { userModel, productModel, orderModel };
