+++ src/middlewares/index.js
@@ -1,4 +1,4 @@
 import { customError, errorHandler } from './error-handler.js';
-import verifyToken from './verifyToken.js';
+import verifyToken from './verify-token.js';
 
 export { customError, errorHandler, verifyToken };
