+++ src/routers/index.js
@@ -1,7 +1,7 @@
-import { userRouter } from './userRouter.js';
-import { orderRouter } from './orderRouter.js';
-import { productRouter } from './productRouter.js';
-import { authRouter } from './authRouter.js';
-import { viewRouter } from './viewRouter.js';
+import { userRouter } from './user-router.js';
+import { orderRouter } from './order-router.js';
+import { productRouter } from './product-router.js';
+import { authRouter } from './auth-router.js';
+import { viewRouter } from './view-router.js';
 
 export { userRouter, orderRouter, productRouter, authRouter, viewRouter };
