+++ src/services/index.js
@@ -1,7 +1,7 @@
-import userService from './userService.js';
-import orderService from './orderService.js';
-import authService from './authService.js';
-import viewService from './viewService.js';
-import productService from './productService.js';
+import userService from './user-service.js';
+import orderService from './order-service.js';
+import authService from './auth-service.js';
+import viewService from './view-service.js';
+import productService from './product-service.js';
 
 export { userService, orderService, authService, viewService, productService };
