+++ src/utils/index.js
@@ -1,4 +1,4 @@
 import httpLogStream from './logger.js';
-import { __filename, __dirname } from './findPath.js';
+import { __filename, __dirname } from './find-path.js';
 
 export { httpLogStream, __filename, __dirname };
